+++ src/utils/const.js
@@ -1,0 +1,24 @@
+export const Films = {
+  TOTAL_AMOUNT: 10,
+  INITIAL_AMOUNT: 5,
+  LOAD_AMOUNT: 5,
+  EXTRA_FILM_AMOUNT: 2
+};
+
+export const userRatings = [
+  {
+    title: `Novice`,
+    min: 0,
+    max: 9
+  },
+  {
+    title: `Fun`,
+    min: 10,
+    max: 19
+  },
+  {
+    title: `Movie Buff`,
+    min: 20,
+    max: 30
+  }
+];
